fix(stylelint-plugin): break out rule logic [KHCP-10186]
