+++ stylelint-plugin/rules/use-proper-token/index.js
@@ -1,11 +1,11 @@
 const stylelint = require('stylelint')
 const { ruleMessages, validateOptions, report } = stylelint.utils
-const PROPERTY_TOKEN_MAP = require('./token-map')
-const { KONG_TOKEN_PREFIX, extractTokensFromValue, RULE_NAME_PREFIX } = require('../../utilities')
+const { RULE_NAME_PREFIX } = require('../../utilities')
+const getInappropriateTokens = require('./getInappropriateTokens')
 
 const ruleName = `${RULE_NAME_PREFIX}/use-proper-token`
 const messages = ruleMessages(ruleName, {
-  unexpected: (token, property) => `Unexpected usage of '${token}' token in '${property}' property.`,
+  unexpected: (token, property) => `Unexpected usage of "${token}" token in "${property}" property`,
 })
 const meta = {
   url: 'https://github.com/Kong/design-tokens/blob/main/stylelint-plugin/README.md',
@@ -21,41 +21,13 @@ const ruleFunction = () => {
     }
 
     postcssRoot.walkDecls((decl) => {
-      /**
-       * PostCSS AST declaration node
-       * Docs: https://postcss.org/api/#declaration
-       */
-
-      const declProp = decl.prop
-      const declValue = decl.value
-      // check if the value contains a token as CSS or SCSS variable
-      const hasToken = [`--${KONG_TOKEN_PREFIX}`, `$${KONG_TOKEN_PREFIX}`].some(pattern => declValue.includes(pattern))
-      if (!hasToken) {
-        // skip validating if the value does not contain a token
-        return
-      }
-
-      const tokenProperty = Object.keys(PROPERTY_TOKEN_MAP).find(key => key.split(',').some(prop => prop === declProp))
-
-      // check if the property is in the property map
-      const isEnforcedProp = !!tokenProperty
-      if (!isEnforcedProp) {
-        return
-      }
-
-      const valueTokens = extractTokensFromValue(declValue)
-      // get the appropriate tokens for the property and create regex for each
-      // regex pattern: /kui-(?:[a-z0-9-]+-)?token(?:-[a-z0-9-]+)?$/
-      // this allows to match both regular and component tokens
-      const appropriateTokens = PROPERTY_TOKEN_MAP[tokenProperty].map(token => new RegExp(KONG_TOKEN_PREFIX + '(?:[a-z0-9-]+-)?' + token + '(?:-[a-z0-9-]+)?$'))
-      // filter out tokens that are not appropriate for the property
-      const inappropriateTokens = valueTokens.filter(vToken => !appropriateTokens.some(aTokenRegex => aTokenRegex.test(vToken)))
+      const inappropriateTokens = getInappropriateTokens(decl)
 
       if (inappropriateTokens.length) {
         // report the error for each inappropriate token
         inappropriateTokens.forEach((token) => {
           report({
-            message: messages.unexpected(token, declProp),
+            message: messages.unexpected(token, decl.prop),
             node: decl,
             result: postcssResult,
             ruleName,
